requirements.txt
ipykernel
langchain
langchain-community
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
pinecone